requirements.txt
pyaudio
pygame
numpy
waveshare-epd
pillow